screenrecord: Add audio support Adds audio recording to the screenrecord system. Note that while recording, the audio is muted from the device output, and that the audio volume is controlled separately (so make sure to turn up your volume if you hear nothing). Audio from the microphone can also be recorded by adding --microphone to the command line.
screenrecord: Enable SUBMIX_IN as well and cleanup on exit

While on Nexus 5/Nexus 7 (AOSP QCOM audio), trying to turn on
SUBMIX_IN fails and is not needed, it is required on Find 5
and N1 (CAF QCOM audio), or otherwise SUBMIX_OUT won't turn
on.
We also now release the audio routing as it wouldn't revive
the audio speakers otherwise.
https://github.com/DirtyUnicorns-KitKat/frameworks_av/commit/09d70217c9e29d5127bfa1216662f5c9fc77366b

Change-Id: If99d4e4c0aba4a453b5cdb8d1225e65d0431d671